Change position of create/clone button
Move the Create button from review.js to navButtons.js
