+++ frontend/src/components/projectCreate/review.js
@@ -1,50 +1,13 @@
-import React, { useState, useCallback } from 'react';
-import { navigate } from '@reach/router';
-import truncate from '@turf/truncate';
-import { FormattedMessage, useIntl } from 'react-intl';
+import React, { useState } from 'react';
+import { FormattedMessage } from 'react-intl';
 
 import messages from './messages';
-import { Button } from '../button';
 import { AlertMessage } from './alertMessage';
-import { createProject } from '../../store/actions/project';
-import { store } from '../../store';
-import { pushToLocalJSONAPI } from '../../network/genericJSONRequest';
+
 import { OrganisationSelect } from '../formInputs';
 
 export default function Review({ metadata, updateMetadata, token, projectId, cloneProjectData }) {
   const [error, setError] = useState(null);
-  const intl = useIntl();
-
-  const handleCreate = useCallback(
-    (metadata, token, cloneProjectData, setError) => {
-      if (!metadata.geom) {
-        setError(intl.formatMessage(messages.noGeometry));
-        return;
-      }
-      if (!metadata.organisation && !cloneProjectData.organisation) {
-        setError(intl.formatMessage(messages.noOrganization));
-        return;
-      }
-
-      store.dispatch(createProject(metadata));
-      let projectParams = {
-        areaOfInterest: truncate(metadata.geom, { precision: 6 }),
-        projectName: metadata.projectName,
-        organisation: metadata.organisation || cloneProjectData.organisation,
-        tasks: truncate(metadata.taskGrid, { precision: 6 }),
-        arbitraryTasks: metadata.arbitraryTasks,
-      };
-
-      if (cloneProjectData.name !== null) {
-        projectParams.projectName = '';
-        projectParams.cloneFromProjectId = cloneProjectData.id;
-      }
-      pushToLocalJSONAPI('projects/', JSON.stringify(projectParams), token)
-        .then((res) => navigate(`/manage/projects/${res.projectId}`))
-        .catch((e) => setError(e.Error));
-    },
-    [intl],
-  );
 
   const setProjectName = (event) => {
     event.preventDefault();
@@ -93,18 +56,6 @@ export default function Review({ metadata, updateMetadata, token, projectId, clo
         </>
       ) : null}
 
-      <div className="mt4">
-        <Button
-          onClick={() => handleCreate(metadata, token, cloneProjectData, setError)}
-          className="white bg-red"
-        >
-          {cloneProjectData.name === null ? (
-            <FormattedMessage {...messages.create} />
-          ) : (
-            <FormattedMessage {...messages.clone} />
-          )}
-        </Button>
-      </div>
       {error && (
         <div className="mt3 red">
           <AlertMessage
